feat: 게시판 CRUD 기능 + swagger jwt 인증 추가

+++ controller/Cboard.js
@@ -16,104 +16,193 @@ const { token } = require('morgan');
 // 그룹의 공지 게시판
 exports.getGroupNotiBoard = async (req, res) => {
   //   const gSeq = req.query.gSeq; // 클라이언트에서 보낸 gSeq
-  const gSeq = req.params.gSeq;
-  const category = 'notice';
+  try {
+    const gSeq = req.params.gSeq;
+    const category = 'notice';
 
-  console.log(gSeq, category);
-  console.log(req.params); // gSeq값
+    console.log(gSeq, category);
+    console.log(req.params); // gSeq값
 
-  const groupInfo = await GroupBoard.findAll({
-    where: { gSeq: gSeq, gbCategory: category },
-  });
+    const groupInfo = await GroupBoard.findAll({
+      where: { gSeq: gSeq, gbCategory: category },
+    });
+    console.log(groupInfo);
 
-  res.json(groupInfo);
+    if (groupInfo) {
+      // 게시글을 찾았을 경우
+      res.status(200).send({
+        success: true,
+        msg: '게시글 조회 성공',
+        groupInfo,
+      });
+    } else {
+      // 게시글을 찾지 못했을 경우
+      res.status(404).send({
+        success: false,
+        msg: '게시글을 찾을 수 없습니다.',
+      });
+    }
+  } catch {
+    res.status(405).send({
+      success: false,
+      msg: 'gSeq를 찾을 수 없습니다.',
+    });
+  }
 };
 
 // 그룹의 공지 게시글 디테알
 exports.getGroupNotiDetail = async (req, res) => {
-  const gSeq = req.params.gSeq;
-  const category = 'notice'; // 'notice' 카테고리
-  const gbSeq = req.params.gbSeq; // 게시글 고유 식별자
+  try {
+    const gSeq = req.params.gSeq;
+    const category = 'notice'; // 'notice' 카테고리
+    const gbSeq = req.params.gbSeq; // 게시글 고유 식별자
 
-  console.log(gSeq, '그룹의', gbSeq, '번째 글, 카테고리는', category);
-  console.log(req.params); // gSeq값
+    console.log(gSeq, '그룹의', gbSeq, '번째 글, 카테고리는', category);
+    console.log(req.params); // gSeq값
 
-  // gSeq 및 category로 공지사항 게시글 필터링
-  const groupInfo = await GroupBoard.findOne({
-    where: { gSeq: gSeq, gbCategory: category, gbSeq: gbSeq },
-  });
+    // gSeq 및 category로 공지사항 게시글 필터링
+    const groupInfo = await GroupBoard.findOne({
+      where: { gSeq: gSeq, gbCategory: category, gbSeq: gbSeq },
+    });
 
-  if (groupInfo) {
-    // 게시글을 찾았을 경우
-    res.status(200).json(groupInfo);
-  } else {
-    // 게시글을 찾지 못했을 경우
-    res.status(404).json({ error: '게시글을 찾을 수 없습니다.' });
+    if (groupInfo) {
+      // 게시글을 찾았을 경우
+      res.status(200).send({
+        success: true,
+        msg: '게시글 조회 성공',
+        groupInfo,
+      });
+    } else {
+      // 게시글을 찾지 못했을 경우
+      res.status(404).send({
+        success: false,
+        msg: '게시글을 찾을 수 없습니다.',
+      });
+    }
+  } catch {
+    res.status(405).send({
+      success: false,
+      msg: 'gSeq 혹은 gbSeq를 찾을 수 없습니다.',
+    });
   }
 };
 
 // 그룹의 자유 게시판
 exports.getGroupFreeBoard = async (req, res) => {
-  const gSeq = req.params.gSeq;
-  const category = 'free';
+  try {
+    const gSeq = req.params.gSeq;
+    const category = 'free';
 
-  console.log(gSeq, category);
-  console.log(req.params); // gSeq값
+    console.log(gSeq, category);
+    console.log(req.params); // gSeq값
 
-  const groupInfo = await GroupBoard.findAll({
-    where: { gSeq: gSeq, gbCategory: category },
-  });
+    const groupInfo = await GroupBoard.findAll({
+      where: { gSeq: gSeq, gbCategory: category },
+    });
 
-  res.json(groupInfo);
+    if (groupInfo) {
+      // 게시글을 찾았을 경우
+      res.status(200).send({
+        success: true,
+        msg: '게시글 조회 성공',
+        groupInfo,
+      });
+    } else {
+      // 게시글을 찾지 못했을 경우
+      res.status(404).send({
+        success: false,
+        msg: '게시글을 찾을 수 없습니다.',
+      });
+    }
+  } catch {
+    res.status(405).send({
+      success: false,
+      msg: 'gSeq를 찾을 수 없습니다.',
+    });
+  }
 };
 
 // 그룹의 자유 게시글 디테알
 exports.getGroupFreeDetail = async (req, res) => {
-  const gSeq = req.params.gSeq;
-  const category = 'free'; // 'free' 카테고리
-  const gbSeq = req.params.gbSeq; // 게시글 고유 식별자
+  try {
+    const gSeq = req.params.gSeq;
+    const category = 'free'; // 'free' 카테고리
+    const gbSeq = req.params.gbSeq; // 게시글 고유 식별자
 
-  console.log(gSeq, '그룹의', gbSeq, '번째 글, 카테고리는', category);
-  console.log(req.params); // gSeq값
+    console.log(gSeq, '그룹의', gbSeq, '번째 글, 카테고리는', category);
+    console.log(req.params); // gSeq값
 
-  // gSeq 및 category로 공지사항 게시글 필터링
-  const groupInfo = await GroupBoard.findOne({
-    where: { gSeq: gSeq, gbCategory: category, gbSeq: gbSeq },
-  });
+    // gSeq 및 category로 공지사항 게시글 필터링
+    const groupInfo = await GroupBoard.findOne({
+      where: { gSeq: gSeq, gbCategory: category, gbSeq: gbSeq },
+    });
 
-  if (groupInfo) {
-    // 게시글을 찾았을 경우
-    res.status(200).json(groupInfo);
-  } else {
-    // 게시글을 찾지 못했을 경우
-    res.status(404).json({ error: '게시글을 찾을 수 없습니다.' });
+    if (groupInfo) {
+      // 게시글을 찾았을 경우
+      res.status(200).send({
+        success: true,
+        msg: '게시글 조회 성공',
+        groupInfo,
+      });
+    } else {
+      // 게시글을 찾지 못했을 경우
+      res.status(404).send({
+        success: false,
+        msg: '게시글을 찾을 수 없습니다.',
+      });
+    }
+  } catch {
+    res.status(405).send({
+      success: false,
+      msg: 'gSeq 혹은 gbSeq 를 찾을 수 없습니다.',
+    });
   }
 };
 
 // 그룹의 미션 게시판
 exports.getGroupMissionBoard = async (req, res) => {
-  const gSeq = req.params.gSeq;
-  const category = 'mission';
-  const mSeq = req.query.mSeq; // 클라이언트에서 요청 보낼때 query로 mSeq 값 넣어서 보내주기
+  try {
+    const gSeq = req.params.gSeq;
+    const category = 'mission';
+    const mSeq = req.params.mSeq; // 클라이언트에서 요청 보낼때 query로 mSeq 값 넣어서 보내주기
 
-  console.log(gSeq, category, mSeq);
+    console.log(gSeq, category, mSeq);
 
-  const groupInfo = await GroupBoard.findAll({
-    where: { gSeq: gSeq, gbCategory: category, mSeq: mSeq },
-  });
+    const groupInfo = await GroupBoard.findAll({
+      where: { gSeq: gSeq, gbCategory: category, mSeq: mSeq },
+    });
 
-  res.json(groupInfo);
+    if (groupInfo) {
+      // 게시글을 찾았을 경우
+      res.status(200).send({
+        success: true,
+        msg: '게시글 조회 성공',
+        groupInfo,
+      });
+    } else {
+      // 게시글을 찾지 못했을 경우
+      res.status(404).send({
+        success: false,
+        msg: '게시글을 찾을 수 없습니다.',
+      });
+    }
+  } catch {
+    res.status(405).send({
+      success: false,
+      msg: 'gSeq 혹은 mSeq를 찾을 수 없습니다.',
+    });
+  }
 };
 
 // 새 게시글 생성 페이지 렌더링
 // /board/create
 exports.getCreateBoard = async (req, res) => {
-  let token = req.headers.authorization.split(' ')[1];
-  const user = await jwt.verify(token);
-  console.log('디코딩 된 토큰!!!!!!!!!!! :', user);
+  // let token = req.headers.authorization.split(' ')[1];
+  // const user = await jwt.verify(token);
+  // console.log('디코딩 된 토큰!!!!!!!!!!! :', user);
 
-  const uSeq = user.uSeq;
-  console.log(uSeq);
+  // const uSeq = user.uSeq;
+  // console.log(uSeq);
 
   // // 클라이언트에서 요청 보낼때 query로 mSeq, gSeq, category 값 넣어서 보내주기
   // const mSeq = req.query.mSeq;
@@ -121,8 +210,21 @@ exports.getCreateBoard = async (req, res) => {
   // const category = req.query.category;
 
   try {
-    if (!uSeq) {
+    let token = req.headers.authorization.split(' ')[1];
+    const user = await jwt.verify(token);
+    console.log('디코딩 된 토큰!!!!!!!!!!! :', user);
+
+    const uSeq = user.uSeq;
+    console.log(uSeq);
+
+    if (!token) {
       res.status(401).send({
+        success: false,
+        msg: '토큰 X',
+      });
+    }
+    if (!uSeq) {
+      res.status(402).send({
         success: false,
         msg: '로그인X or 비정상적인 접근',
       });
@@ -131,7 +233,8 @@ exports.getCreateBoard = async (req, res) => {
 
     res.status(200).send({
       success: true,
-      msg: '게시글 생성 페이지 로딩 성공',
+      msg: '게시글 조회 성공',
+      groupInfo,
       data: {
         user: uSeq,
       },
@@ -141,7 +244,7 @@ exports.getCreateBoard = async (req, res) => {
     console.log(error);
     res.status(500).send({
       success: false,
-      msg: '서버에러 발생',
+      msg: '서버 에러',
     });
   }
 };
@@ -150,27 +253,43 @@ exports.getCreateBoard = async (req, res) => {
 // /board/create
 exports.createBoard = async (req, res) => {
   //   authUtil.checkToken();
-  let token = req.headers.authorization.split(' ')[1];
-  //   console.log('=========', req.headers, token);
-  const user = await jwt.verify(token);
-  console.log('디코딩 된 토큰!!!!!!!!!!! :', user);
+  // let token = req.headers.authorization.split(' ')[1];
+  // //   console.log('=========', req.headers, token);
+  // const user = await jwt.verify(token);
+  // console.log('디코딩 된 토큰!!!!!!!!!!! :', user);
 
-  const uSeq = user.uSeq;
-  console.log(uSeq);
-
-  // 클라이언트에서 요청 보낼때 query로 mSeq, gSeq, category 값 넣어서 보내주기
-  const gSeq = req.query.gSeq;
-  const category = req.query.category;
+  // const uSeq = user.uSeq;
+  // console.log(uSeq);
 
   try {
-    console.log('gSeq : ', gSeq);
-    if (!uSeq) {
+    let token = req.headers.authorization.split(' ')[1];
+    const user = await jwt.verify(token);
+    console.log('디코딩 된 토큰!!!!!!!!!!! :', user);
+
+    const uSeq = user.uSeq;
+    console.log(uSeq);
+
+    // 클라이언트에서 요청 보낼때 query로 mSeq, gSeq, category 값 넣어서 보내주기
+    const gSeq = req.query.gSeq;
+    const category = req.query.category;
+
+    if (!token) {
       res.status(401).send({
+        success: false,
+        msg: '토큰 X',
+      });
+    }
+
+    if (!uSeq) {
+      res.status(402).send({
         success: false,
         msg: '로그인X or 비정상적인 접근',
       });
       return;
     }
+
+    console.log('gSeq : ', gSeq);
+
     // 모임원인지 아닌지 guSeq 확인
     const groupUser = await GroupUser.findOne({
       where: { uSeq: uSeq },
@@ -250,13 +369,29 @@ const hasChanged = (beforeEdit, afterEdit) =>
 // 게시글 수정 페이지 렌더링
 // /board/edit/:gbSeq
 exports.getEditBoard = async (req, res) => {
-  let token = req.headers.authorization.split(' ')[1];
-  const user = await jwt.verify(token);
-  console.log('디코딩 된 토큰!!!!!!!!!!! :', user);
-
-  const uSeq = user.uSeq;
-  console.log(uSeq);
   try {
+    let token = req.headers.authorization.split(' ')[1];
+    const user = await jwt.verify(token);
+    console.log('디코딩 된 토큰!!!!!!!!!!! :', user);
+
+    const uSeq = user.uSeq;
+    console.log(uSeq);
+
+    if (!token) {
+      res.status(401).send({
+        success: false,
+        msg: '토큰 X',
+      });
+    }
+
+    if (!uSeq) {
+      res.status(402).send({
+        success: false,
+        msg: '로그인X or 비정상적인 접근',
+      });
+      return;
+    }
+
     // 업데이트 전 게시글 데이터 조회
     const gbSeq = req.params.gbSeq;
 
@@ -294,14 +429,35 @@ exports.getEditBoard = async (req, res) => {
 // 게시글 수정 요청
 // /board/edit/:gbSeq
 exports.editBoard = async (req, res) => {
-  let token = req.headers.authorization.split(' ')[1];
-  const user = await jwt.verify(token);
-  console.log('디코딩 된 토큰!!!!!!!!!!! :', user);
+  // let token = req.headers.authorization.split(' ')[1];
+  // const user = await jwt.verify(token);
+  // console.log('디코딩 된 토큰!!!!!!!!!!! :', user);
 
-  const uSeq = user.uSeq;
-  console.log(uSeq);
+  // const uSeq = user.uSeq;
+  // console.log(uSeq);
 
   try {
+    let token = req.headers.authorization.split(' ')[1];
+    const user = await jwt.verify(token);
+    console.log('디코딩 된 토큰!!!!!!!!!!! :', user);
+
+    const uSeq = user.uSeq;
+    console.log(uSeq);
+
+    if (!token) {
+      res.status(401).send({
+        success: false,
+        msg: '토큰 X',
+      });
+    }
+
+    if (!uSeq) {
+      res.status(402).send({
+        success: false,
+        msg: '로그인X or 비정상적인 접근',
+      });
+      return;
+    }
     // 업데이트 전 게시글 데이터 조회
     const gbSeq = req.params.gbSeq;
     // 업데이트 전 게시글 데이터 조회
@@ -357,16 +513,40 @@ exports.editBoard = async (req, res) => {
 // 게시글 삭제 처리
 // /board/delete/:gbSeq
 exports.deleteBoard = async (req, res) => {
-  let token = req.headers.authorization.split(' ')[1];
-  const user = await jwt.verify(token);
-  console.log('디코딩 된 토큰!!!!!!!!!!! :', user);
+  // let token = req.headers.authorization.split(' ')[1];
+  // const user = await jwt.verify(token);
+  // console.log('디코딩 된 토큰!!!!!!!!!!! :', user);
 
-  const uSeq = user.uSeq;
-  console.log(uSeq);
+  // const uSeq = user.uSeq;
+  // console.log(uSeq);
 
-  const gbSeq = req.params.gbSeq;
+  // const gbSeq = req.params.gbSeq;
 
   try {
+    let token = req.headers.authorization.split(' ')[1];
+    const user = await jwt.verify(token);
+    console.log('디코딩 된 토큰!!!!!!!!!!! :', user);
+
+    const uSeq = user.uSeq;
+    console.log(uSeq);
+
+    if (!token) {
+      res.status(401).send({
+        success: false,
+        msg: '토큰 X',
+      });
+    }
+
+    if (!uSeq) {
+      res.status(402).send({
+        success: false,
+        msg: '로그인X or 비정상적인 접근',
+      });
+      return;
+    }
+    const gbSeq = req.params.gbSeq;
+    console.log('삭제하려는 gbSeq : ', gbSeq);
+
     const isDeleted = await GroupBoard.destroy({
       where: {
         gbSeq: gbSeq,
@@ -377,14 +557,14 @@ exports.deleteBoard = async (req, res) => {
     if (!isDeleted) {
       // 삭제 실패 처리
       res.status(404).send({
-        isDeleted: false,
+        success: false,
         msg: '게시글이 삭제되지 않았습니다.',
       });
       return;
     } else {
       // 정상 삭제 처리
       res.status(200).send({
-        isDeleted: true,
+        success: true,
         msg: '게시글이 정상적으로 삭제되었습니다.',
       });
     }
@@ -392,43 +572,8 @@ exports.deleteBoard = async (req, res) => {
     console.log(error);
     // 에러 처리
     res.status(500).send({
+      success: false,
       msg: '게시글 삭제처리 중 서버에러 발생',
     });
   }
 };
-
-// // 토큰 검증 함수
-// exports.isAvailableToken = async (req) => {
-//   // let token = req.headers.authorization.split(' ')[1];
-//   // //   console.log('=========', req.headers, token);
-//   // const user = await jwt.verify(token);
-//   // console.log('디코딩 된 토큰!!!!!!!!!!! :', user);
-
-//   // const uSeq = user.uSeq;
-//   // console.log(uSeq);
-//   let token = req.headers.authorization.split(' ')[1];
-
-//   if (!token) {
-//     return { error: '토큰 없음' };
-//   }
-
-//   const user = await jwt.verify(token);
-//   console.log('디코딩 된 토큰입니다 이거슨 :', user);
-
-//   const uSeq = user.uSeq;
-//   console.log(uSeq);
-
-//   if (user === TOKEN_EXPIRED) {
-//     return { error: '유효기간 만료된 토큰' };
-//   }
-
-//   if (user === TOKEN_INVALID) {
-//     return { error: '유효하지 않는 토큰' };
-//   }
-
-//   if (user.uSeq === undefined) {
-//     return { error: '유효하지 않는 토큰: undefined' };
-//   }
-
-//   return { user };
-// };
